Move display related things from hover to display
